+++ CHANGELOG.md
@@ -16,8 +16,10 @@ Not yet released
 
 * Add some additional null checks to services
 * [CS-177][] Fix for untagged images showing as ':' in UI
+* [CS-276][] Correctly track command and wrapper identifiers in LaunchReport
 
 [CS-177]: https://issues.xnat.org/browse/CS-177
+[CS-276]: https://issues.xnat.org/browse/CS-276
 
 ## 1.0-alpha.3-hotfix
 [Released 2017-07-25](https://github.com/NrgXnat/container-service/releases/tag/v1.0-alpha.3-hotfix).